[P2P-199] 댓글 개수 상이 내용 수정

+++ src/main/java/com/prgrms/p2p/domain/place/service/PlaceService.java
@@ -62,7 +62,7 @@ public class PlaceService {
     Place place = placeRepository.findById(placeId)
         .orElseThrow(() -> new NotFoundException("장소가 존재하지 않습니다."));
 
-    Integer commentCount = placeCommentRepository.countByPlace(place);
+    Long commentCount = placeCommentRepository.countByPlace(place);
 
     return toDetailPlaceResponse(place, userId, commentCount);
   }
